Move ChartDataProvider to context folder
